+++ tests/epyccel/test_epyccel_dicts.py
@@ -141,3 +141,43 @@ def test_pop_item_key(language):
     python_result = pop_item()
     assert isinstance(python_result, type(pyccel_result))
     assert python_result == pyccel_result
+
+def test_get_element(language) :
+    def get_element():
+        a = {1:1.0, 2:2.0}
+        return a.get(1)
+    epyc_element = epyccel(get_element, language = language)
+    pyccel_result = epyc_element()
+    python_result = get_element()
+    assert isinstance(python_result, type(pyccel_result))
+    assert python_result == pyccel_result
+
+def test_get_default_element(language) :
+    def get_default_element():
+        a = {1:True, 2:False}
+        return a.get(3, True)
+    epyc_default_element = epyccel(get_default_element, language = language)
+    pyccel_result = epyc_default_element()
+    python_result = get_default_element()
+    assert isinstance(python_result, type(pyccel_result))
+    assert python_result == pyccel_result
+
+def test_get_str_keys(language) :
+    def get_str_keys():
+        a = {'a':1, 'b':2}
+        return a.get('a')
+    epyc_str_keys = epyccel(get_str_keys, language = language)
+    pyccel_result = epyc_str_keys()
+    python_result = get_str_keys()
+    assert isinstance(python_result, type(pyccel_result))
+    assert python_result == pyccel_result
+
+def test_get_default_str_keys(language) :
+    def get_default_str_keys():
+        a = {'a':1, 'b':2}
+        return a.get('c', 4)
+    epyc_str_keys = epyccel(get_default_str_keys, language = language)
+    pyccel_result = epyc_str_keys()
+    python_result = get_default_str_keys()
+    assert isinstance(python_result, type(pyccel_result))
+    assert python_result == pyccel_result